Update to the latest Burn and pin dependencies (#33)
* Add Cargo.lock

The presence of a lock file doesn't affect the versions that other crates
will use, and it ensures that tests in CI are reproducible. Rust currently
recommends not checking it in for libraries, but this appears to be about
to change:
https://github.com/rust-lang/cargo/issues/8728

* Update to the latest Burn; pin to a specific commit

* Silence a clippy warning

https://rust-lang.github.io/rust-clippy/master/index.html#/single_range_in_vec_init
was triggering for lines like

  let rating = ratings.clone().slice([i..i + 1]).transpose();

I think the warning probably doesn't apply for this API.

* Fix a clippy warning

* Pin the Rust version

Allowing it to float will lead to future clippy breakages, as clippy
routinely adds extra checks in new updates.